+++ api/src/main/java/net/infumia/frame/element/item/ElementItemBuilder.java
@@ -1,10 +1,11 @@
-package net.infumia.frame.element;
+package net.infumia.frame.element.item;
 
 import java.util.function.Consumer;
 import java.util.function.Function;
 import net.infumia.frame.context.element.ContextElementItemClick;
 import net.infumia.frame.context.element.ContextElementItemRender;
 import net.infumia.frame.context.element.ContextElementItemUpdate;
+import net.infumia.frame.element.ElementBuilder;
 import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.NotNull;
