+++ api/src/main/java/net/infumia/frame/element/pagination/ElementPaginationBuilder.java
@@ -3,7 +3,7 @@ package net.infumia.frame.element.pagination;
 import java.util.function.BiConsumer;
 import net.infumia.frame.context.ContextBase;
 import net.infumia.frame.element.ElementBuilder;
-import net.infumia.frame.element.ElementItemBuilder;
+import net.infumia.frame.element.item.ElementItemBuilder;
 import net.infumia.frame.state.pagination.ElementConfigurer;
 import net.infumia.frame.state.pagination.StatePagination;
 import org.jetbrains.annotations.NotNull;
