+++ api/src/main/java/net/infumia/frame/state/pagination/ElementConfigurer.java
@@ -1,7 +1,7 @@
 package net.infumia.frame.state.pagination;
 
 import net.infumia.frame.context.ContextBase;
-import net.infumia.frame.element.ElementItemBuilder;
+import net.infumia.frame.element.item.ElementItemBuilder;
 import org.jetbrains.annotations.NotNull;
 
 @FunctionalInterface
