+++ core/src/main/java/net/infumia/frame/context/element/ContextElementItemClickImpl.java
@@ -1,6 +1,6 @@
 package net.infumia.frame.context.element;
 
-import net.infumia.frame.element.ElementItem;
+import net.infumia.frame.element.item.ElementItem;
 import org.jetbrains.annotations.NotNull;
 
 public final class ContextElementItemClickImpl
