+++ core/src/main/java/net/infumia/frame/context/view/ContextRenderImpl.java
@@ -12,9 +12,9 @@ import net.infumia.frame.Preconditions;
 import net.infumia.frame.context.ContextBase;
 import net.infumia.frame.context.ContextBaseImpl;
 import net.infumia.frame.element.Element;
-import net.infumia.frame.element.ElementItemBuilder;
-import net.infumia.frame.element.ElementItemBuilderImpl;
-import net.infumia.frame.element.ElementItemBuilderRich;
+import net.infumia.frame.element.item.ElementItemBuilder;
+import net.infumia.frame.element.item.ElementItemBuilderImpl;
+import net.infumia.frame.element.item.ElementItemBuilderRich;
 import net.infumia.frame.extension.CompletableFutureExtensions;
 import net.infumia.frame.metadata.MetadataAccess;
 import net.infumia.frame.metadata.MetadataKeyHolder;
