+++ core/src/main/java/net/infumia/frame/element/ElementImpl.java
@@ -19,17 +19,17 @@ import org.jetbrains.annotations.Nullable;
 public class ElementImpl implements ElementRich {
 
     private final String key;
-    final ContextBase parent;
-    final Element root;
-    final boolean cancelOnClick;
-    final boolean updateOnClick;
-    final boolean closeOnClick;
-    final Duration interactionDelay;
-    final Consumer<ContextElementClick> onInteractionDelay;
-    final Function<ContextElementClick, String> interactionDelayKey;
-    final Predicate<ContextElementRender> displayIf;
-    final Collection<State<?>> updateOnStateChange;
-    final Collection<State<?>> updateOnStateAccess;
+    protected final ContextBase parent;
+    protected final Element root;
+    protected final boolean cancelOnClick;
+    protected final boolean updateOnClick;
+    protected final boolean closeOnClick;
+    protected final Duration interactionDelay;
+    protected final Consumer<ContextElementClick> onInteractionDelay;
+    protected final Function<ContextElementClick, String> interactionDelayKey;
+    protected final Predicate<ContextElementRender> displayIf;
+    protected final Collection<State<?>> updateOnStateChange;
+    protected final Collection<State<?>> updateOnStateAccess;
     private boolean visible = true;
 
     public ElementImpl(
