--- core/src/main/java/net/infumia/frame/element/ElementItemRich.java
@@ -1,3 +1,0 @@
-package net.infumia.frame.element;
-
-public interface ElementItemRich extends ElementRich, ElementItem, ElementEventHandlerHolder {}
