+++ core/src/main/java/net/infumia/frame/element/item/ElementItemBuilderRich.java
@@ -1,4 +1,6 @@
-package net.infumia.frame.element;
+package net.infumia.frame.element.item;
+
+import net.infumia.frame.element.ElementBuilderRich;
 
 public interface ElementItemBuilderRich extends ElementBuilderRich, ElementItemBuilder {
     int slot();
