+++ core/src/main/java/net/infumia/frame/element/item/ElementItemRich.java
@@ -1,0 +1,6 @@
+package net.infumia.frame.element.item;
+
+import net.infumia.frame.element.ElementEventHandlerHolder;
+import net.infumia.frame.element.ElementRich;
+
+public interface ElementItemRich extends ElementRich, ElementItem, ElementEventHandlerHolder {}
