+++ core/src/main/java/net/infumia/frame/element/pagination/ElementPaginationImpl.java
@@ -18,10 +18,10 @@ import net.infumia.frame.element.Element;
 import net.infumia.frame.element.ElementBuilder;
 import net.infumia.frame.element.ElementEventHandler;
 import net.infumia.frame.element.ElementImpl;
-import net.infumia.frame.element.ElementItem;
-import net.infumia.frame.element.ElementItemBuilderImpl;
-import net.infumia.frame.element.ElementItemBuilderRich;
 import net.infumia.frame.element.ElementRich;
+import net.infumia.frame.element.item.ElementItem;
+import net.infumia.frame.element.item.ElementItemBuilderImpl;
+import net.infumia.frame.element.item.ElementItemBuilderRich;
 import net.infumia.frame.extension.CompletableFutureExtensions;
 import net.infumia.frame.pipeline.executor.PipelineExecutorElement;
 import net.infumia.frame.pipeline.executor.PipelineExecutorElementImpl;
