+++ core/src/main/java/net/infumia/frame/pipeline/service/render/ServiceFirstRenderAvailableSlotResolution.java
@@ -6,7 +6,7 @@ import java.util.function.BiFunction;
 import net.infumia.frame.context.view.ContextRender;
 import net.infumia.frame.context.view.ContextRenderRich;
 import net.infumia.frame.element.ElementBuilderRich;
-import net.infumia.frame.element.ElementItemBuilder;
+import net.infumia.frame.element.item.ElementItemBuilder;
 import net.infumia.frame.pipeline.PipelineServiceConsumer;
 import net.infumia.frame.pipeline.context.PipelineContextRender;
 import net.infumia.frame.slot.LayoutSlot;
