+++ core/src/main/java/net/infumia/frame/pipeline/service/render/ServiceFirstRenderConsumeNonRenderedElement.java
@@ -2,7 +2,7 @@ package net.infumia.frame.pipeline.service.render;
 
 import java.util.Iterator;
 import net.infumia.frame.context.view.ContextRenderRich;
-import net.infumia.frame.element.ElementItemBuilderRich;
+import net.infumia.frame.element.item.ElementItemBuilderRich;
 import net.infumia.frame.pipeline.PipelineServiceConsumer;
 import net.infumia.frame.pipeline.context.PipelineContextRender;
 import org.jetbrains.annotations.NotNull;
