+++ core/src/main/java/net/infumia/frame/slot/SlotFinder.java
@@ -9,8 +9,8 @@ import net.infumia.frame.InvTypeRich;
 import net.infumia.frame.Preconditions;
 import net.infumia.frame.SlotConverter;
 import net.infumia.frame.context.view.ContextRender;
-import net.infumia.frame.element.ElementItemBuilder;
-import net.infumia.frame.element.ElementItemBuilderRich;
+import net.infumia.frame.element.item.ElementItemBuilder;
+import net.infumia.frame.element.item.ElementItemBuilderRich;
 import net.infumia.frame.view.ViewContainerRich;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
